+++ packages/modernjs/src/cli/ssrPlugin.ts
@@ -7,13 +7,15 @@ import { ModuleFederationPlugin as RspackModuleFederationPlugin } from '@module-
 import UniverseEntryChunkTrackerPlugin from '@module-federation/node/universe-entry-chunk-tracker-plugin';
 import { updateStatsAndManifest } from './manifest';
 import { isDev } from './constant';
+import { MODERN_JS_SERVER_DIR } from '../constant';
 import logger from './logger';
+import { isWebTarget } from './utils';
 
 export function setEnv() {
   process.env['MF_DISABLE_EMIT_STATS'] = 'true';
   process.env['MF_SSR_PRJ'] = 'true';
 }
-
+export const CHAIN_MF_PLUGIN_ID = 'plugin-module-federation-server';
 export const moduleFederationSSRPlugin = (
   pluginOptions: Required<InternalModernPluginOptions>,
 ): CliPluginFuture<AppTools> => ({
@@ -24,9 +26,13 @@ export const moduleFederationSSRPlugin = (
   ],
   setup: async (api) => {
     const modernjsConfig = api.getConfig();
-    const enableSSR = Boolean(modernjsConfig?.server?.ssr);
-    if (!enableSSR || pluginOptions.userConfig?.ssr === false) {
-      return {} as any;
+    const enableSSR =
+      pluginOptions.userConfig?.ssr ?? Boolean(modernjsConfig?.server?.ssr);
+    let csrOutputPath = '';
+    let ssrOutputPath = '';
+
+    if (!enableSSR) {
+      return;
     }
 
     setEnv();
@@ -42,44 +48,55 @@ export const moduleFederationSSRPlugin = (
       });
       return { entrypoint, plugins };
     });
+    api.modifyBundlerChain((chain, { isServer }) => {
+      const bundlerType =
+        api.getAppContext().bundlerType === 'rspack' ? 'rspack' : 'webpack';
+      const MFPlugin =
+        bundlerType === 'webpack'
+          ? ModuleFederationPlugin
+          : RspackModuleFederationPlugin;
+
+      const isWeb = isWebTarget(chain.get('target'));
+
+      if (!isWeb) {
+        if (!chain.plugins.has(CHAIN_MF_PLUGIN_ID)) {
+          chain
+            .plugin(CHAIN_MF_PLUGIN_ID)
+            .use(MFPlugin, [pluginOptions.ssrConfig])
+            .init((Plugin: typeof MFPlugin, args) => {
+              pluginOptions.nodePlugin = new Plugin(args[0]);
+              return pluginOptions.nodePlugin;
+            });
+        }
+      }
+
+      if (!isWeb) {
+        chain.target('async-node');
+        if (isDev) {
+          chain
+            .plugin('UniverseEntryChunkTrackerPlugin')
+            .use(UniverseEntryChunkTrackerPlugin);
+        }
+      }
+
+      if (isDev && !isServer) {
+        chain.externals({
+          '@module-federation/node/utils': 'NOT_USED_IN_BROWSER',
+        });
+      }
+
+      if (isServer) {
+        ssrOutputPath =
+          chain.output.get('path') ||
+          path.resolve(process.cwd(), `dist/${MODERN_JS_SERVER_DIR}`);
+      } else {
+        csrOutputPath =
+          chain.output.get('path') || path.resolve(process.cwd(), 'dist');
+      }
+    });
     api.config(() => {
       return {
         tools: {
-          rspack(config, { isServer }) {
-            if (isServer) {
-              // throw new Error(
-              //   `${PLUGIN_IDENTIFIER} Not support rspack ssr mode yet !`,
-              // );
-              if (!pluginOptions.nodePlugin) {
-                pluginOptions.nodePlugin = new RspackModuleFederationPlugin(
-                  pluginOptions.ssrConfig,
-                );
-                // @ts-ignore
-                config.plugins?.push(pluginOptions.nodePlugin);
-              }
-            } else {
-              pluginOptions.distOutputDir =
-                pluginOptions.distOutputDir ||
-                config.output?.path ||
-                path.resolve(process.cwd(), 'dist');
-            }
-          },
-          webpack(config, { isServer }) {
-            if (isServer) {
-              if (!pluginOptions.nodePlugin) {
-                pluginOptions.nodePlugin = new ModuleFederationPlugin(
-                  pluginOptions.ssrConfig,
-                );
-                // @ts-ignore
-                config.plugins?.push(pluginOptions.nodePlugin);
-              }
-            } else {
-              pluginOptions.distOutputDir =
-                pluginOptions.distOutputDir ||
-                config.output?.path ||
-                path.resolve(process.cwd(), 'dist');
-            }
-          },
           devServer: {
             before: [
               (req, res, next) => {
@@ -111,31 +128,26 @@ export const moduleFederationSSRPlugin = (
               },
             ],
           },
-          bundlerChain(chain, { isServer }) {
-            if (isServer) {
-              chain.target('async-node');
-              if (isDev) {
-                chain
-                  .plugin('UniverseEntryChunkTrackerPlugin')
-                  .use(UniverseEntryChunkTrackerPlugin);
-              }
-            }
-            if (isDev && !isServer) {
-              chain.externals({
-                '@module-federation/node/utils': 'NOT_USED_IN_BROWSER',
-              });
-            }
-          },
         },
       };
     });
     api.onAfterBuild(() => {
       const { nodePlugin, browserPlugin, distOutputDir } = pluginOptions;
-      updateStatsAndManifest(nodePlugin, browserPlugin, distOutputDir);
+      updateStatsAndManifest(
+        nodePlugin,
+        browserPlugin,
+        distOutputDir,
+        path.relative(csrOutputPath, ssrOutputPath),
+      );
     });
     api.onDevCompileDone(() => {
       const { nodePlugin, browserPlugin, distOutputDir } = pluginOptions;
-      updateStatsAndManifest(nodePlugin, browserPlugin, distOutputDir);
+      updateStatsAndManifest(
+        nodePlugin,
+        browserPlugin,
+        distOutputDir,
+        path.relative(csrOutputPath, ssrOutputPath),
+      );
     });
   },
 });