+++ packages/rspack/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @module-federation/rspack
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies [0f71cbc]
+- Updated dependencies [22fcccd]
+  - @module-federation/sdk@0.10.0
+  - @module-federation/dts-plugin@0.10.0
+  - @module-federation/bridge-react-webpack-plugin@0.10.0
+  - @module-federation/managers@0.10.0
+  - @module-federation/manifest@0.10.0
+  - @module-federation/runtime-tools@0.10.0
+  - @module-federation/inject-external-runtime-core-plugin@0.10.0
+
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies [35d925b]
+- Updated dependencies [35d925b]
+- Updated dependencies [7a37360]
+- Updated dependencies [8acd217]
+  - @module-federation/dts-plugin@0.9.1
+  - @module-federation/sdk@0.9.1
+  - @module-federation/manifest@0.9.1
+  - @module-federation/bridge-react-webpack-plugin@0.9.1
+  - @module-federation/managers@0.9.1
+  - @module-federation/runtime-tools@0.9.1
+  - @module-federation/inject-external-runtime-core-plugin@0.9.1
+
+## 0.9.0
+
+### Patch Changes
+
+- Updated dependencies [5ebc53f]
+- Updated dependencies [d872b35]
+  - @module-federation/dts-plugin@0.9.0
+  - @module-federation/manifest@0.9.0
+  - @module-federation/sdk@0.9.0
+  - @module-federation/runtime-tools@0.9.0
+  - @module-federation/managers@0.9.0
+  - @module-federation/bridge-react-webpack-plugin@0.9.0
+  - @module-federation/inject-external-runtime-core-plugin@0.9.0
+
+## 0.8.12
+
+### Patch Changes
+
+- Updated dependencies [d227303]
+  - @module-federation/dts-plugin@0.8.12
+  - @module-federation/manifest@0.8.12
+  - @module-federation/sdk@0.8.12
+  - @module-federation/runtime-tools@0.8.12
+  - @module-federation/managers@0.8.12
+  - @module-federation/bridge-react-webpack-plugin@0.8.12
+  - @module-federation/inject-external-runtime-core-plugin@0.8.12
+
+## 0.8.11
+
+### Patch Changes
+
+- @module-federation/sdk@0.8.11
+- @module-federation/runtime-tools@0.8.11
+- @module-federation/managers@0.8.11
+- @module-federation/manifest@0.8.11
+- @module-federation/dts-plugin@0.8.11
+- @module-federation/bridge-react-webpack-plugin@0.8.11
+- @module-federation/inject-external-runtime-core-plugin@0.8.11
+
+## 0.8.10
+
+### Patch Changes
+
+- @module-federation/dts-plugin@0.8.10
+- @module-federation/runtime-tools@0.8.10
+- @module-federation/inject-external-runtime-core-plugin@0.8.10
+- @module-federation/sdk@0.8.10
+- @module-federation/managers@0.8.10
+- @module-federation/manifest@0.8.10
+- @module-federation/bridge-react-webpack-plugin@0.8.10
+
 ## 0.8.9
 
 ### Patch Changes